Intreset Calculator
In this project, I developed an interest calculator using Python and Tkinter. The calculator is designed to compute both simple and compound interest. It features a clear, user-friendly interface with buttons to easily switch between the two types of calculations. The Tkinter library was used to build the graphical interface, while Python functions handle the mathematical computations. This makes it a versatile tool for quickly and accurately managing interest calculations.