Text to Speech
This project is a Text-to-Speech app built with Python and Tkinter. Users can input text and choose the voice (Male or Female) and speed (Slow, Normal, Fast) using dropdown menus. The app provides two main functions: click "Speech" to hear the text read aloud, and click "Save" to save the speech as an MP3 file. The interface also includes a "Quit" button to close the app. The design features a straightforward layout with clear labels and buttons for ease of use.