Spelling Corrector
This project is a Spelling Corrector application created with Python and Tkinter. It uses the TextBlob library to correct spelling errors in the input text. The app features a text area where users can enter words. After clicking the "Check" button, the application uses TextBlob to correct the spelling and displays the corrected word. The interface also includes a "Quit" button to close the application. The design is simple, with a clear layout and easy-to-read fonts.