ChatBot
This project is a basic chatbot app built with Python’s Tkinter. Users interact with the chatbot, Aditya, through a graphical interface. The bot replies to predefined messages and provides a time-based greeting. Users type messages into an entry field, and both user and bot messages appear in a scrollable text area. A "Send" button submits messages and updates the conversation display