Display Data
This project is a web application that displays data from a database. Users can enter a specific ID to retrieve and view details of a particular user in a table format. Additionally, the application provides the option to view all records from the database, offering a comprehensive overview of the stored information. The application is built with Flask and utilizes dynamic templates to present data effectively and interactively..