Universe Game
This project is a simple game where the user must protect themselves from various challenges or obstacles to achieve a high score. The gameplay focuses on survival and strategic movement, with the primary objective being to maximize the score by successfully navigating and responding to in-game threats. The game is designed to be engaging and challenging, providing an enjoyable experience as users strive to improve their performance and reach new high scores.