Number to Words
This project is a number-to-words converter application. Users input a number, and the application converts it into its textual representation. Built with Flask, the app uses request to capture user input, render_template to present the results, and the num2words library to perform the conversion. This setup provides an easy-to-use interface for translating numerical values into written words.