ASCII art Generator
This project is a text-to-ASCII art converter. Users can input any text, and the application transforms it into visually engaging ASCII art. Built with Python and utilizing the pyfiglet library, the app provides a straightforward interface for converting standard text into a variety of artistic ASCII styles. This tool is ideal for adding creative text decorations or unique visual elements to any text-based content.