Emojii Creator
This project is an emoji generator that converts words into corresponding emojis. Users enter a word, and the application returns a relevant emoji representation of that word. Built with Python and utilizing the emoji library, the app provides a simple interface for enhancing text with visual emoji elements, making communication more expressive and engaging.