Email Sender
This project is an email sender application with a user interface modeled after Google Mail (Gmail). Built with Python and smtplib, it allows users to compose and send emails via an SMTP server. The UI replicates the look and feel of Gmail, offering a familiar and intuitive experience for managing email communication.